refactor: move out function dublicates to own file
